Renamed GLDemo to GLWidget
Still trying to figure out good naming policies for this stuff. A "demo"
is a demonstration of something. In our case, it always contains some
OpenGL widget. In many current cases, the demo *is* the widget.

To apply this policy, I needed to rename many areas that I called
"demos" since they're actually "widgets."