Refactored rotation methods, making ROTATION_SCALE static
In preparation for merging logo code back into GLWidget, I needed to
add a new set of rotation setters that expect scaled values. These allow
us to have intuitive syntax for children, as is done in SpiralGLWidget,
as well as having sliders be able to pass scaled values directly, as is
done in LogoGLWidget.