rotation must be initialized to 0, added initializer for timer
In the C++ world, values aren't necessarily initialized to 0; in this
case, rotation would get initialized to nan. This would cause any
rotation to fail.

I fixed it by using C++'s initializer syntax. It's a bit more efficient
than putting it in the constructor; it would be required if these things
were const.