AnimatedGLWidget properly uses signal/slot
This behavior is similar to the pure virtual case. However, we emit a
timeout(const float&) signal that we attach in our constructor. The
attached slot is a pure virtual. This way, we get the easy behavior of a
pure virtual function, along with the flexibility of the signal down the
road.