Pulled out even more animation stuff into MeasuredTimer
This separates timer management from GLWidget-specific behavior. While
this is more lines of code, I think this produces two easy-to-understand
classes. It also allows this animation code to be used without needing
to inherit a GLWidget class.

AnimatedGLWidget's logic is rather trivial now that MeasuredTimer
exists. I'm not ready to remove it yet, since it is convenient to have
around. It also demonstrates the minimum code necessary to use
MeasuredTimer.