Vector3 now has a wrapping policy
This enables us to wrap vector's values without need to make set{XYZ}
virtual, which is a decent performance gain for this low-level class.
Ordinary use of Vector3 remains unchanged, but if you want the
components of Vector3 to be wrapped, you can do so by creating it as
follows:

	Vector3<float, NormalizeWrap> vec;

	vec.addX(3 * M_PI);
	assert(vec.x() == M_PI);

There will likely never be way to change the wrapping policy of an
individual vector run-time. I think this makes sense: It would be
unusual to wrap at 2pi, then suddenly wrap at pi.