Use const char* in Player, rather than just char*
This was causing a warning. The reason this should be const is because
the values passed is a string literal, which can't be safely modified.

There's also a different between a pointer to const data and a const
pointer to modifiable data. The first case means we can change what we
point to, but we can't modify the underlying data. The second case means
the data can be modified, but the pointer itself can't be changed.

First case (const data, modifiable pointer) : const char*
Second case (modifiable data, const pointer): char* const

There's nothing stopping us from a const pointer to const data, which
would be:

Third case (const data, const pointer): const char* const