RenderLayer: Inlined position template param
RenderLayers are buried under several layers in a fully constructed
pipeline, and none of those classes are suitable for providing the
template param directly. As a result, keeping the param would require
all the layered classes to expose it themselves. This caused some very
annoying code to be generated, for what is really an implementation
detail in RenderLayer itself.

I used double because it's the highest-precision portable data type.
This is suitable for a camera, which relatively few instances of will
exist at a given time.