VoxampRenderLayer: Adjust the transformed pos'n of rendered voxels
The origin used for rendering voxels is now always at the center of that
voxel. Since the "direction" in which a voxel should be rendered may
change depending on the render layer's configuration, this would require
voxel renderers to require the direction as well. However, if we render
the voxel at its center, there's no need to provide a direction.