DirectRenderLayer: Renderer is now a std::function
This significantly increases the flexibility of a direct render layer.
The price is that renderables are now keyed by their physical alone. It
is not possible, during removal, to distinguish between two Renderables
that have the same Physical.

To demonstrate, consider the following situation:

void renderOneWay();
void renderAnotherWay();

Physical<float> myObject;

DirectRenderLayer<float> layer;
layer.add(&myObject, renderOneWay);
layer.add(&myObject, renderAnotherWay);

// ... later ...

// Both renderOneWay and renderAnotherWay are removed!
layer.remove(&myObject);

If this limitation becomes significant, we can add an integer ID to
allow the removal of Renderables. However, for now, this problem can be
easily avoided, and in general it seems relatively unlikely to occur.